+++ src/room/track/RemoteAudioTrack.ts
@@ -1,10 +1,11 @@
 import log from '../../logger';
 import { TrackEvent } from '../events';
-import { AudioReceiverStats, computeBitrate } from '../stats';
+import { computeBitrate } from '../stats';
+import type { AudioReceiverStats } from '../stats';
 import { supportsSetSinkId } from '../utils';
-import type { AudioOutputOptions } from './options';
 import RemoteTrack from './RemoteTrack';
 import { Track } from './Track';
+import type { AudioOutputOptions } from './options';
 
 export default class RemoteAudioTrack extends RemoteTrack {
   private prevStats?: AudioReceiverStats;